+++ Libraries/LibCore/CProcessStatisticsReader.cpp
@@ -19,7 +19,7 @@ HashMap<pid_t, CProcessStatistics> CProcessStatisticsReader::get_all()
     HashMap<pid_t, CProcessStatistics> map;
 
     auto file_contents = file->read_all();
-    auto json = JsonValue::from_string({ file_contents.data(), file_contents.size() });
+    auto json = JsonValue::from_string({ file_contents.data(), (size_t)file_contents.size() });
     json.as_array().for_each([&](auto& value) {
         const JsonObject& process_object = value.as_object();
         CProcessStatistics process;
